backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
aiofiles==24.1.0
openai==1.59.9
pydantic==2.10.5
pydantic-settings==2.7.1
python-dotenv==1.0.1
anthropic>=0.40.0
